rename generateData to generate-data
